Added missing lyrics methods for dashes and extends
