+++ src/MusicalScore/Graphical/StaffLine.ts
@@ -8,6 +8,7 @@ import {StaffMeasure} from "./StaffMeasure";
 import {MusicSystem} from "./MusicSystem";
 import {StaffLineActivitySymbol} from "./StaffLineActivitySymbol";
 import {PointF2D} from "../../Common/DataObjects/PointF2D";
+import {GraphicalLabel} from "./GraphicalLabel";
 
 /**
  * A StaffLine contains the [[Measure]]s in one line of the music sheet
@@ -21,6 +22,7 @@ export abstract class StaffLine extends GraphicalObject {
     protected skyLine: number[];
     protected bottomLine: number[];
     protected lyricLines: GraphicalLine[] = [];
+    protected lyricsDashes: GraphicalLabel[] = [];
 
     constructor(parentSystem: MusicSystem, parentStaff: Staff) {
         super();
@@ -51,6 +53,14 @@ export abstract class StaffLine extends GraphicalObject {
 
     public set LyricLines(value: GraphicalLine[]) {
         this.lyricLines = value;
+    }
+
+    public get LyricsDashes(): GraphicalLabel[] {
+        return this.lyricsDashes;
+    }
+
+    public set LyricsDashes(value: GraphicalLabel[]) {
+        this.lyricsDashes = value;
     }
 
     public get ParentMusicSystem(): MusicSystem {
